gemini/function2/requirements.txt
flask
functions-framework==3.*
google-cloud-storage

google-cloud-aiplatform
google-cloud-texttospeech
google-genai
